Added data validation and error handling

+++ controllers/shrines.js
@@ -21,11 +21,17 @@ const getShrine = async (req, res) => {
     #swagger.summary = 'Get a shrine by ID'
     */
     try {
+        if (!ObjectId.isValid(req.params.id)) {
+            return res.status(400).json({ message: 'Invalid Shrine ID' });
+        };
         const result = await mongodb.getDb().db().collection('shrines').findOne({ _id: new ObjectId(req.params.id) });
+        if (!result) {
+            return res.status(404).json({ message: 'Shrine not found' });
+        };
         res.setHeader('Content-Type', 'application/json');
         res.status(200).json(result);
     } catch (error) {
-        res.status(500).json(error);
+        res.status(500).json(error );
     }
 };
 
@@ -45,7 +51,7 @@ const createShrine = async (req, res) => {
         res.setHeader('Content-Type', 'application/json');
         res.status(201).json(result);
     } catch (error) {
-        res.status(500).json(error);
+        res.status(500).json(error || 'An error occurred while creating the shrine');
     }
 };
 
@@ -60,12 +66,18 @@ const updateShrine = async (req, res) => {
     }
     */
     try {
+        if (!ObjectId.isValid(req.params.id)) {
+            return res.status(400).json({ message: 'Invalid Shrine ID' });
+        }
         const updatedShrine = req.body;
         const result = await mongodb.getDb().db().collection('shrines').updateOne({ _id: new ObjectId(req.params.id) }, { $set: updatedShrine });
+        if (result.matchedCount == 0) {
+            return res.status(404).json({ message: 'Shrine not found' });
+        };
         res.setHeader('Content-Type', 'application/json');
         res.status(204).json(result);
     } catch (error) {
-        res.status(500).json(error);
+        res.status(500).json(error || 'An error occurred while updating the shrine');
     }
 };
 
@@ -74,11 +86,17 @@ const deleteShrine = async (req, res) => {
     #swagger.summary = 'Delete a shrine'
     */
     try {
+        if (!ObjectId.isValid(req.params.id)) {
+            return res.status(400).json({ message: 'Invalid Shrine ID' });
+        }
         const result = await mongodb.getDb().db().collection('shrines').deleteOne({ _id: new ObjectId(req.params.id) });
+        if (result.deletedCount == 0) {
+            return res.status(404).json({ message: 'Shrine not found' });
+        };
         res.setHeader('Content-Type', 'application/json');
         res.status(200).json(result);
     } catch (error) {
-        res.status(500).json(error);
+        res.status(500).json(error || 'An error occurred while deleting the shrine' );
     }
 };
